+++ tests/stubs/helix/commands.scm
@@ -1,6 +1,8 @@
 ;; Stub of helix's virtual helix/commands.scm module, for `make compile-check`.
 ;; See ../README.md.
 
-(provide debug-start)
+(provide debug-start
+         write)
 
 (define (debug-start template . arguments) void)
+(define (write . arguments) void)